Trade related changes

+++ ForexAPITester/CloseTradeRequest.cs
@@ -12,7 +12,7 @@ namespace ForexAPITester
         public string epic { get; set; }
         public string expiry { get; set; }
         public string direction { get; set; }
-        public float size { get; set; }
+        public double size { get; set; }
         public string level { get; set; }
         public string orderType { get; set; } = "MARKET";
         public object timeInForce { get; set; }
